+++ src/photon_platform/clerk/__init__.py
@@ -11,5 +11,5 @@ PHOTON platform
 __author__ = "PHOTON platform"
 __maintainer__ = "PHOTON platform"
 __email__ = "github@phiarchitect.com"
-__version__ = "0.0.1"
+__version__ = "0.0.2"
 __licence__ = "MIT"